feat(Breadcrumb): retrieve items in Story
